compitable for runing sed under macos

+++ scripts/local-purge.sh
@@ -1,8 +1,10 @@
 #!/usr/bin/env bash
 set -e
 
-docker stop $(docker ps -f name=local_ -a -q)
-docker rm $(docker ps -f name=local_ -a -q)
-docker volume rm $(docker volume ls -q -f dangling=true -f name=local_)
-
-sed -i '/dev: /d' .graphcoolrc
+containers=$(docker ps -f name=local_ -a -q)
+if [ ! -z "$containers" ]; then
+  docker stop $containers
+  docker rm $containers
+  docker volume rm $(docker volume ls -q -f dangling=true -f name=local_)
+  sed -i'.bak' '/dev: /d' .graphcoolrc
+fi
